+++ jason/service.py
@@ -1,4 +1,5 @@
 import threading
+import flask
 from typing import Iterable
 
 from jason.core import configuration
@@ -24,8 +25,8 @@ class Service:
             thread.start()
         try:
             self.main()
-        except Exception:
-            self.on_error()
+        except Exception as ex:
+            self.on_error(ex)
             raise
         else:
             self.tear_down()
@@ -36,13 +37,26 @@ class Service:
         ...
 
     def main(self):
-        raise NotImplementedError
-
-    def on_error(self):
         ...
 
     def tear_down(self):
         ...
 
+    def on_error(self, ex):
+        ...
+
     def on_close(self):
         ...
+
+
+class FlaskService(Service):
+
+    def __init__(self, *args, **kwargs):
+        super(FlaskService, self).__init__(*args, **kwargs)
+        self.app = self.create_app()
+
+    def create_app(self):
+        app = flask.Flask(__name__)
+        app.config.update(self.config.__dict__)
+        app.config["DB_HANDLER"] = self.db_handler
+        return app