+++ RELEASES.md
@@ -3,6 +3,8 @@ v0.0.8
 - Renamed core.config to core.configuration
 - Added config module for config mixins
 - Added postgres config mixin with a couple of helper functions
+- Added service module
+- Added database module
 
 v0.0.7
 ===